requirements.txt
streamlit>=1.32.0
groq>=0.9.0
pdfplumber>=0.10.3
python-docx>=1.1.0
pandas>=2.2.0
nltk>=3.8.1
